+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-  "liveServer.settings.port": 5502,
+  "liveServer.settings.port": 5503,
   "githubPullRequests.ignoredPullRequestBranches": [
     "main"
   ]
