+++ index.html
@@ -16,8 +16,7 @@
         <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;600&family=Roboto&family=Source+Serif+4:opsz,wght@8..60,700&display=swap" rel="stylesheet">
         <link rel="stylesheet" href="./css/index.css">
         <script src="https://kit.fontawesome.com/2c5c2a6dff.js" crossorigin="anonymous"></script>
-        <script type="module" src="./scripts/index.mjs"></script>
-        <script defer src="./scripts/cart.js"></script>
+        <script type="module" src="./scripts/display-films.mjs"></script>
     </head>
     <body>
         <header>
